requirements.txt
# API и сервер
fastapi
uvicorn
pydantic
jinja2
requests

# Работа с моделями
torch
torchvision
sentence-transformers
transformers
faiss-cpu
evaluate
huggingface_hub

# Работа с данными
numpy
pandas
scikit-learn
datasets

# Визуализация и логи
matplotlib
seaborn
tqdm
wandb

# Работа с аргументами командной строки
argparse

# Файловые операции, регулярки и прочее (не требуют установки)
shutil
re
warnings
